+++ src/NetBox/WinSCPPlugin.h
@@ -51,7 +51,7 @@ protected:
   virtual intptr_t ProcessEditorInputEx(const INPUT_RECORD * Rec);
   bool CopyParamDialog(const UnicodeString & Caption, TCopyParamType & CopyParam,
     int CopyParamAttrs);
-  void MessageClick(void * Token, int Result, bool & Close);
+  void MessageClick(void * Token, intptr_t Result, bool & Close);
 
   void CommandsMenu(bool FromFileSystem);
   bool ConfigurationDialog();
